inline collections_controller_resource_spec to existing tests
`dashboard/collections_controller_resource_spec` is a slightly divergent
copy-paste of the same tests in `dashboard/collections_controller_spec`. this
creates a major maintenance burden, and some weird misdirection when navigating
the codebase (how will i know to look for tests on the Controller here?).

the more immediate problem is that this whole test setup is flaky and the whole
file can fail en masse.

as a first step(?) for a fix, we want to inline all the tests so they run
identically and setup for both sets of specs is in one place. this is a straight
inline, removing the duplicate tests. some attention was given to resolving
drift between the specs.
